refactor(spec) put test schemas in namespaces
