+++ spec/graphql/analysis/field_usage_spec.rb
@@ -5,7 +5,7 @@ describe GraphQL::Analysis::FieldUsage do
   let(:result) { [] }
   let(:field_usage_analyzer) { GraphQL::Analysis::FieldUsage.new { |query, used_fields, used_deprecated_fields| result << query << used_fields << used_deprecated_fields } }
   let(:reduce_result) { GraphQL::Analysis.analyze_query(query, [field_usage_analyzer]) }
-  let(:query) { GraphQL::Query.new(DummySchema, query_string, variables: variables) }
+  let(:query) { GraphQL::Query.new(Dummy::Schema, query_string, variables: variables) }
   let(:variables) { {} }
 
   describe "query with deprecated fields" do
