+++ spec/graphql/directive_spec.rb
@@ -3,7 +3,7 @@ require "spec_helper"
 
 describe GraphQL::Directive do
   let(:variables) { {"t" => true, "f" => false} }
-  let(:result) { DummySchema.execute(query_string, variables: variables) }
+  let(:result) { Dummy::Schema.execute(query_string, variables: variables) }
   describe "on fields" do
     let(:query_string) { %|query directives($t: Boolean!, $f: Boolean!) {
       cheese(id: 1) {
