+++ spec/graphql/execution/typecast_spec.rb
@@ -5,7 +5,7 @@ describe GraphQL::Execution::Typecast do
   let(:milk_value) { MILKS[1] }
   let(:cheese_value) { CHEESES[1] }
 
-  let(:schema) { DummySchema }
+  let(:schema) { Dummy::Schema }
   let(:context) { GraphQL::Query::Context.new(query: OpenStruct.new(schema: schema), values: nil) }
 
   def compatible?(*args)
@@ -13,40 +13,40 @@ describe GraphQL::Execution::Typecast do
   end
 
   it "resolves correctly when both types are the same" do
-    assert compatible?(MilkType, MilkType, context)
+    assert compatible?(Dummy::MilkType, Dummy::MilkType, context)
 
-    assert !compatible?(MilkType, CheeseType, context)
+    assert !compatible?(Dummy::MilkType, Dummy::CheeseType, context)
   end
 
   it "resolves a union type to a matching member" do
-    assert compatible?(DairyProductUnion, MilkType, context)
-    assert compatible?(DairyProductUnion, CheeseType, context)
+    assert compatible?(Dummy::DairyProductUnion, Dummy::MilkType, context)
+    assert compatible?(Dummy::DairyProductUnion, Dummy::CheeseType, context)
 
-    assert !compatible?(DairyProductUnion, GraphQL::INT_TYPE, context)
-    assert !compatible?(DairyProductUnion, HoneyType, context)
+    assert !compatible?(Dummy::DairyProductUnion, GraphQL::INT_TYPE, context)
+    assert !compatible?(Dummy::DairyProductUnion, Dummy::HoneyType, context)
   end
 
   it "resolves correcty when potential type is UnionType and current type is a member of that union" do
-    assert compatible?(MilkType, DairyProductUnion, context)
-    assert compatible?(CheeseType, DairyProductUnion, context)
+    assert compatible?(Dummy::MilkType, Dummy::DairyProductUnion, context)
+    assert compatible?(Dummy::CheeseType, Dummy::DairyProductUnion, context)
 
-    assert !compatible?(QueryType, DairyProductUnion, context)
-    assert !compatible?(EdibleInterface, DairyProductUnion, context)
+    assert !compatible?(Dummy::DairyAppQueryType, Dummy::DairyProductUnion, context)
+    assert !compatible?(Dummy::EdibleInterface, Dummy::DairyProductUnion, context)
   end
 
   it "resolves an object type to one of its interfaces" do
-    assert compatible?(CheeseType, EdibleInterface, context)
-    assert compatible?(MilkType, EdibleInterface, context)
+    assert compatible?(Dummy::CheeseType, Dummy::EdibleInterface, context)
+    assert compatible?(Dummy::MilkType, Dummy::EdibleInterface, context)
 
-    assert !compatible?(QueryType, EdibleInterface, context)
-    assert !compatible?(LocalProductInterface, EdibleInterface, context)
+    assert !compatible?(Dummy::DairyAppQueryType, Dummy::EdibleInterface, context)
+    assert !compatible?(Dummy::LocalProductInterface, Dummy::EdibleInterface, context)
   end
 
   it "resolves an interface to a matching member" do
-    assert compatible?(EdibleInterface, CheeseType, context)
-    assert compatible?(EdibleInterface, MilkType, context)
+    assert compatible?(Dummy::EdibleInterface, Dummy::CheeseType, context)
+    assert compatible?(Dummy::EdibleInterface, Dummy::MilkType, context)
 
-    assert !compatible?(EdibleInterface, GraphQL::STRING_TYPE, context)
-    assert !compatible?(EdibleInterface, DairyProductInputType, context)
+    assert !compatible?(Dummy::EdibleInterface, GraphQL::STRING_TYPE, context)
+    assert !compatible?(Dummy::EdibleInterface, Dummy::DairyProductInputType, context)
   end
 end
