+++ spec/graphql/internal_representation/rewrite_spec.rb
@@ -2,8 +2,8 @@
 require "spec_helper"
 
 describe GraphQL::InternalRepresentation::Rewrite do
-  let(:validator) { GraphQL::StaticValidation::Validator.new(schema: DummySchema) }
-  let(:query) { GraphQL::Query.new(DummySchema, query_string) }
+  let(:validator) { GraphQL::StaticValidation::Validator.new(schema: Dummy::Schema) }
+  let(:query) { GraphQL::Query.new(Dummy::Schema, query_string) }
   let(:rewrite_result) {
     validator.validate(query)[:irep]
   }
@@ -24,18 +24,18 @@ describe GraphQL::InternalRepresentation::Rewrite do
     it "produces a tree of nodes" do
       op_node = rewrite_result["getCheeses"]
 
-      root_children = op_node.typed_children[DairyAppQueryType]
+      root_children = op_node.typed_children[Dummy::DairyAppQueryType]
       assert_equal 2, root_children.length
-      assert_equal DairyAppQueryType, op_node.return_type
+      assert_equal Dummy::DairyAppQueryType, op_node.return_type
       first_field = root_children.values.first
-      assert_equal 3, first_field.typed_children[CheeseType].length
-      assert_equal DairyAppQueryType, first_field.owner_type
-      assert_equal CheeseType, first_field.return_type
+      assert_equal 3, first_field.typed_children[Dummy::CheeseType].length
+      assert_equal Dummy::DairyAppQueryType, first_field.owner_type
+      assert_equal Dummy::CheeseType, first_field.return_type
 
       second_field = root_children.values.last
-      assert_equal 1, second_field.typed_children[CheeseType].length
-      assert_equal DairyAppQueryType.get_field("cheese"), second_field.definition
-      assert_equal CheeseType, second_field.return_type
+      assert_equal 1, second_field.typed_children[Dummy::CheeseType].length
+      assert_equal Dummy::DairyAppQueryType.get_field("cheese"), second_field.definition
+      assert_equal Dummy::CheeseType, second_field.return_type
       assert second_field.inspect.is_a?(String)
     end
   end
@@ -50,8 +50,8 @@ describe GraphQL::InternalRepresentation::Rewrite do
     |}
 
     it "gets dynamic field definitions" do
-      cheese_field = rewrite_result[nil].typed_children[DairyAppQueryType]["cheese"]
-      typename_field = cheese_field.typed_children[CheeseType]["typename"]
+      cheese_field = rewrite_result[nil].typed_children[Dummy::DairyAppQueryType]["cheese"]
+      typename_field = cheese_field.typed_children[Dummy::CheeseType]["typename"]
       assert_equal "__typename", typename_field.definition.name
       assert_equal "__typename", typename_field.definition_name
     end
@@ -128,41 +128,41 @@ describe GraphQL::InternalRepresentation::Rewrite do
     it "puts all fragment members as children" do
       op_node = rewrite_result[nil]
 
-      cheese_field = op_node.typed_children[DairyAppQueryType]["cheese"]
-      assert_equal ["id1", "id2", "fatContent", "similarCow", "flavor"], cheese_field.typed_children[CheeseType].keys
-      assert_equal ["fatContent", "origin"], cheese_field.typed_children[EdibleInterface].keys
+      cheese_field = op_node.typed_children[Dummy::DairyAppQueryType]["cheese"]
+      assert_equal ["id1", "id2", "fatContent", "similarCow", "flavor"], cheese_field.typed_children[Dummy::CheeseType].keys
+      assert_equal ["fatContent", "origin"], cheese_field.typed_children[Dummy::EdibleInterface].keys
       # Merge:
-      similar_cow_field = cheese_field.typed_children[CheeseType]["similarCow"]
-      assert_equal ["similarCowSource", "fatContent", "similarCheese", "id"], similar_cow_field.typed_children[CheeseType].keys
+      similar_cow_field = cheese_field.typed_children[Dummy::CheeseType]["similarCow"]
+      assert_equal ["similarCowSource", "fatContent", "similarCheese", "id"], similar_cow_field.typed_children[Dummy::CheeseType].keys
       # Deep merge:
-      similar_sheep_field = similar_cow_field.typed_children[CheeseType]["similarCheese"]
-      assert_equal ["flavor", "source"], similar_sheep_field.typed_children[CheeseType].keys
+      similar_sheep_field = similar_cow_field.typed_children[Dummy::CheeseType]["similarCheese"]
+      assert_equal ["flavor", "source"], similar_sheep_field.typed_children[Dummy::CheeseType].keys
 
-      edible_origin_node = cheese_field.typed_children[EdibleInterface]["origin"]
-      assert_equal EdibleInterface.get_field("origin"), edible_origin_node.definition
-      assert_equal EdibleInterface, edible_origin_node.owner_type
+      edible_origin_node = cheese_field.typed_children[Dummy::EdibleInterface]["origin"]
+      assert_equal Dummy::EdibleInterface.get_field("origin"), edible_origin_node.definition
+      assert_equal Dummy::EdibleInterface, edible_origin_node.owner_type
 
-      edible_fat_content_node = cheese_field.typed_children[EdibleInterface]["fatContent"]
-      assert_equal EdibleInterface.get_field("fatContent"), edible_fat_content_node.definition
-      assert_equal EdibleInterface, edible_fat_content_node.owner_type
+      edible_fat_content_node = cheese_field.typed_children[Dummy::EdibleInterface]["fatContent"]
+      assert_equal Dummy::EdibleInterface.get_field("fatContent"), edible_fat_content_node.definition
+      assert_equal Dummy::EdibleInterface, edible_fat_content_node.owner_type
 
-      cheese_fat_content_node = cheese_field.typed_children[CheeseType]["fatContent"]
-      assert_equal CheeseType.get_field("fatContent"), cheese_fat_content_node.definition
-      assert_equal CheeseType, cheese_fat_content_node.owner_type
+      cheese_fat_content_node = cheese_field.typed_children[Dummy::CheeseType]["fatContent"]
+      assert_equal Dummy::CheeseType.get_field("fatContent"), cheese_fat_content_node.definition
+      assert_equal Dummy::CheeseType, cheese_fat_content_node.owner_type
 
-      cheese_flavor_node = cheese_field.typed_children[CheeseType]["flavor"]
-      assert_equal CheeseType.get_field("flavor"), cheese_flavor_node.definition
-      assert_equal CheeseType, cheese_flavor_node.owner_type
+      cheese_flavor_node = cheese_field.typed_children[Dummy::CheeseType]["flavor"]
+      assert_equal Dummy::CheeseType.get_field("flavor"), cheese_flavor_node.definition
+      assert_equal Dummy::CheeseType, cheese_flavor_node.owner_type
 
 
       # nested spread inside fragment definition:
-      cheese_2_field = op_node.typed_children[DairyAppQueryType]["cheese2"].typed_children[CheeseType]["similarCheese"]
-      assert_equal ["id", "fatContent"], cheese_2_field.typed_children[CheeseType].keys
+      cheese_2_field = op_node.typed_children[Dummy::DairyAppQueryType]["cheese2"].typed_children[Dummy::CheeseType]["similarCheese"]
+      assert_equal ["id", "fatContent"], cheese_2_field.typed_children[Dummy::CheeseType].keys
     end
   end
 
   describe "nested fields on typed fragments" do
-    let(:result) { DummySchema.execute(query_string) }
+    let(:result) { Dummy::Schema.execute(query_string) }
     let(:query_string) {%|
     {
       allDairy {
