+++ spec/graphql/introspection/type_type_spec.rb
@@ -12,7 +12,7 @@ describe GraphQL::Introspection::TypeType do
        missingType:   __type(name: "NotAType") { name }
      }
   |}
-  let(:result) { DummySchema.execute(query_string, context: {}, variables: {"cheeseId" => 2}) }
+  let(:result) { Dummy::Schema.execute(query_string, context: {}, variables: {"cheeseId" => 2}) }
   let(:cheese_fields) {[
     {"name"=>"deeplyNullableCheese", "isDeprecated" => false, "type"=>{ "kind" => "OBJECT", "name" => "Cheese", "ofType" => nil}},
     {"name"=>"flavor",      "isDeprecated" => false, "type" => { "kind" => "NON_NULL", "name" => nil, "ofType" => { "name" => "String"}}},
