+++ spec/graphql/relay/connection_type_spec.rb
@@ -31,7 +31,7 @@ describe GraphQL::Relay::ConnectionType do
         assert_equal 'basesWithCustomEdge', bases["fieldName"]
         assert_equal ["YAVIN", "ECHO BASE", "SECRET HIDEOUT"] , bases["edges"].map { |e| e["upcasedName"] }
         assert_equal ["Yavin", "Echo Base", "Secret Hideout"] , bases["edges"].map { |e| e["node"]["name"] }
-        assert_equal ["CustomBaseEdge"] , bases["edges"].map { |e| e["edgeClassName"] }.uniq
+        assert_equal ["StarWars::CustomBaseEdge"] , bases["edges"].map { |e| e["edgeClassName"] }.uniq
         upcased_rebels_name = "ALLIANCE TO RESTORE THE REPUBLIC"
         assert_equal [upcased_rebels_name] , bases["edges"].map { |e| e["upcasedParentName"] }.uniq
       end
