+++ spec/graphql/static_validation/validator_spec.rb
@@ -2,8 +2,8 @@
 require "spec_helper"
 
 describe GraphQL::StaticValidation::Validator do
-  let(:validator) { GraphQL::StaticValidation::Validator.new(schema: DummySchema) }
-  let(:query) { GraphQL::Query.new(DummySchema, query_string) }
+  let(:validator) { GraphQL::StaticValidation::Validator.new(schema: Dummy::Schema) }
+  let(:query) { GraphQL::Query.new(Dummy::Schema, query_string) }
   let(:errors) { validator.validate(query)[:errors].map(&:to_h) }
