+++ spec/graphql/union_type_spec.rb
@@ -27,7 +27,7 @@ describe GraphQL::UnionType do
   end
 
   describe "typecasting from union to union" do
-    let(:result) { DummySchema.execute(query_string) }
+    let(:result) { Dummy::Schema.execute(query_string) }
     let(:query_string) {%|
       {
         allDairy {
@@ -55,7 +55,7 @@ describe GraphQL::UnionType do
 
   describe "list of union type" do
     describe "fragment spreads" do
-      let(:result) { DummySchema.execute(query_string) }
+      let(:result) { Dummy::Schema.execute(query_string) }
       let(:query_string) {%|
         {
           allDairy {
